Create a ContactCard component and use in Contacts.
